Forward final versions
New Service: motion-format
refactors the format functions from motion-line-numbering into an own
service, streamlines the access of ViewUnifiedObjects to make it easier
to access combined changed versions for every other function
