+++ client/src/app/site/motions/services/motion-format.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MotionFormatService } from './motion-format.service';
+
+describe('MotionFormatService', () => {
+    let service: MotionFormatService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({});
+        service = TestBed.inject(MotionFormatService);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});
